backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.8.6
duckduckgo-search==4.1.1
# Cloud deployment
psycopg[binary]>=3.1.0
cloudinary==1.36.0
